requirements.txt
numpy==1.23.5
opencv-python==4.8.0.76
Pillow==9.5.0
ultralytics==8.0.20